Fix git branch for edk2 builds
Running `./build --package` currently fails with:

  git checkout current/edk2-stable202502
  error: pathspec 'current/edk2-stable202502' did not match any file(s) known to git
  ERROR: git checkout current/edk2-stable202502

This is because stable-commits was updated to point to an edk2 stable
tag, but the scripts don't currently handle tags properly. Instead,
point to snp-latest branch for edk2, which is already updated to be a
mirror of the latest (at this time) edk2 stable tag.

Signed-off-by: Michael Roth <michael.roth@amd.com>